test(phantomjs): Add PhantomJS console-error checking for demo modules
Closes #437

Previously we had a feature to have PhantomJS load the demo page and
ensure that there were no Javascript errors on the console. It was only
loading the main page though, and none of the pages for our components.
This PR now fires off individual PhantomJS tasks for each of the demo
modules.

Note that I had to move all of the module configuration stuff (i.e.
`grunt.config('config.modules')` into a separate task. Previously
it was done in the `build` task, but because of how our Travis is
configured, I needed to be able to grab that configuration info
from outside of `build`, to pass it to the PhantomJS tests.